requirements.txt
Flask==2.2.3
firebase-admin==6.1.0
pycryptodome==3.18.0
pyotp==2.8.0
stripe==5.6.0
bcrypt==4.0.1
requests==2.31.0
